fix bitmap pack filter
