+++ dbms/src/Storages/DeltaMerge/File/DMFilePackFilter.cpp
@@ -22,6 +22,10 @@
 #include <Storages/DeltaMerge/RowKeyRange.h>
 #include <Storages/DeltaMerge/ScanContext.h>
 
+#include <memory>
+
+#include "Storages/DeltaMerge/Index/RSResult.h"
+
 namespace DB::DM
 {
 
@@ -299,6 +303,81 @@ void DMFilePackFilter::tryLoadIndex(RSCheckParam & param, ColId col_id)
 
     Stopwatch watch;
     loadIndex(param.indexes, dmfile, file_provider, index_cache, set_cache_if_miss, col_id, read_limiter, scan_context);
+}
+
+std::pair<std::vector<DMFilePackFilter::Range>, DMFilePackFilterResults> DMFilePackFilter::
+    getSkippedRangeAndFilterForBitmapFilter(
+        const DMContext & dm_context,
+        const DMFiles & dmfiles,
+        const DMFilePackFilterResults & pack_filter_results,
+        UInt64 start_ts)
+{
+    // Packs that all rows compliant with MVCC filter and RowKey filter requirements.
+    // For building bitmap filter, we don't need to read these packs,
+    // just set corresponding positions in the bitmap to true.
+    // So we record the offset and rows of these packs and merge continuous ranges.
+    std::vector<Range> skipped_ranges;
+    // Packs that some rows compliant with MVCC filter and RowKey filter requirements.
+    // We need to read these packs and do RowKey filter and MVCC filter for them.
+    DMFilePackFilterResults new_pack_filter_results;
+    new_pack_filter_results.reserve(dmfiles.size());
+
+    // The offset of the first row in the current range.
+    size_t offset = 0;
+    // The number of rows in the current range.
+    size_t rows = 0;
+    UInt32 preceded_rows = 0;
+
+    auto file_provider = dm_context.global_context.getFileProvider();
+    for (size_t i = 0; i < dmfiles.size(); ++i)
+    {
+        const auto & dmfile = dmfiles[i];
+        const auto & pack_filter = pack_filter_results[i];
+        const auto & pack_res = pack_filter->getPackRes();
+        const auto & handle_res = pack_filter->getHandleRes();
+        const auto & pack_stats = dmfile->getPackStats();
+        DMFilePackFilterResultPtr new_pack_filter;
+        for (size_t pack_id = 0; pack_id < pack_stats.size(); ++pack_id)
+        {
+            const auto & pack_stat = pack_stats[pack_id];
+            preceded_rows += pack_stat.rows;
+            if (!pack_res[pack_id].isUse())
+                continue;
+
+            if (handle_res[pack_id] == RSResult::Some || pack_stat.not_clean > 0
+                || pack_filter->getMaxVersion(dmfile, pack_id, file_provider, dm_context.scan_context) > start_ts)
+            {
+                // We need to read this pack to do RowKey or MVCC filter.
+                continue;
+            }
+
+            if unlikely (!new_pack_filter)
+                new_pack_filter = std::make_shared<DMFilePackFilterResult>(*pack_filter);
+
+            new_pack_filter->pack_res[pack_id] = RSResult::None;
+            // When this pack is next to the previous pack, we merge them.
+            // Otherwise, we record the previous continuous packs and start a new one.
+            if (offset + rows == preceded_rows - pack_stat.rows)
+            {
+                rows += pack_stat.rows;
+            }
+            else
+            {
+                skipped_ranges.emplace_back(offset, rows);
+                offset = preceded_rows - pack_stat.rows;
+                rows = pack_stat.rows;
+            }
+        }
+
+        if (new_pack_filter)
+            new_pack_filter_results.emplace_back(std::move(new_pack_filter));
+        else
+            new_pack_filter_results.emplace_back(pack_filter);
+    }
+    if (rows > 0)
+        skipped_ranges.emplace_back(offset, rows);
+
+    return {skipped_ranges, new_pack_filter_results};
 }
 
 } // namespace DB::DM
